move get changes to openapi utilities
